@@ -28,8 +28,8 @@
     "@babel/preset-env": "^7.14.5",
     "babel-jest": "^27.0.2",
     "babel-loader": "^8.2.2",
-    "css-loader": "^5.2.6",
-    "eslint": "^7.29.0",
+    "css-loader": "^5.2.7",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.4",
@@ -59,6 +59,11 @@
   },
   "keywords": [],
   "dependencies": {
-    "axios": "^0.21.1"
+    "@babel/plugin-transform-runtime": "^7.14.5",
+    "axios": "^0.21.1",
+    "html-loader": "^2.1.2",
+    "sass-loader": "^12.1.0",
+    "style-loader": "^3.2.1",
+    "url-loader": "^4.1.1"
   }
 }
