@@ -1,6 +1,6 @@
 export default class App {
   constructor($target) {
     this.$target = $target;
-    $target.innerText = 'Hello world!';
+    $target.innerText = 'Hello asdffsds!';
   }
 }
